move shutdown signal to util
